requirements.txt
Django==6.0.4
psycopg2==2.9.11
python-decouple==3.8
